+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 0.8.1 February 20, 2013
 Bugfixes and typos
 * Adjust preventDefault behavior for up/down and modifiers
+* Added support for jsonp requests
 * Fix typos
 
 ### 0.8.0 February 19, 2013